+++ doc/source/doxygen-docs/changelog.md
@@ -4,6 +4,10 @@
 - Changes in libraries:
   - \ref mrpt_gui_grp
     - New function mrpt::gui::GetScaledClientSize().
+  - \ref mrpt_obs_grp
+    - Fix typo in data field: mrpt::obs::CObservationRange: `sensorConeApperture`  -> `sensorConeAperture`
+  - \ref mrpt_ros2bridge_grp
+    - Address the new field `variance` in `sensor_msgs/Range` (Closes [issue #1270](https://github.com/MRPT/mrpt/issues/1270)).
 - BUG FIXES:
   - Fix wrong rendering of all wxWidgets-based OpenGL windows when using Ubuntu's display settings to change UI to a size different than 100% (Fixes [issue #1114](https://github.com/MRPT/mrpt/issues/1114)).